+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -23,4 +23,10 @@
 * Changed 
   * Moved core documentation to terra-dev-site.
   * Moved Getting started documentation to terra-dev-site.
-
+  * Updated Bubble graph documentation and examples to be carbon centric.
+  * Updated Line graph documentation and examples to be carbon-centric.
+  * Updated Combination graph documentation and examples to be carbon-centric.
+  * Updated Gantt chart documentation and examples to be carbon-centric.
+  * Updated general graph documentation and examples to be carbon-centric.
+  * Updated Paired Result documentation and examples to be carbon-centric.
+  * Updated Pie chart documentation and examples to be carbon-centric..